+++ requirements.txt
@@ -1,0 +1,4 @@
+numpy==1.24.4
+scipy==1.10.1
+sparse-ir==1.0.0
+xprec==1.3.11
